requirements.txt
elasticsearch==8.10.1
fastapi>=0.111.0
huggingface_hub>=0.22.2
numpy>=1.26.0
protobuf>=4.25.3
psycopg2-binary>=2.9.9
PyMuPDF>=1.24.0
sentence-transformers>=2.7.0
sentencepiece>=0.2.0
tiktoken>=0.7.0
torch>=2.2.0
tqdm>=4.66.0
transformers>=4.40.0,<5
uvicorn>=0.30.0
